Issue 4676: (PDP-34) Initial Implementation (Part 3 of 4) - Move all TableBasedMetadataStore related tests to TableBasedMetadataStoreTests.
