+++ segmentstore/storage/src/test/java/io/pravega/segmentstore/storage/metadata/TableBasedMetadataStoreTests.java
@@ -9,10 +9,17 @@
  */
 package io.pravega.segmentstore.storage.metadata;
 
+import io.pravega.segmentstore.contracts.tables.TableStore;
+import io.pravega.segmentstore.storage.chunklayer.ChunkStorage;
+import io.pravega.segmentstore.storage.chunklayer.ChunkedRollingStorageTests;
+import io.pravega.segmentstore.storage.chunklayer.ChunkedSegmentStorageTests;
+import io.pravega.segmentstore.storage.chunklayer.SimpleStorageTests;
+import io.pravega.segmentstore.storage.mocks.InMemoryChunkStorage;
 import io.pravega.segmentstore.storage.mocks.InMemoryTableStore;
 import lombok.val;
 import org.junit.Before;
 
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
 /**
@@ -23,5 +30,97 @@ public class TableBasedMetadataStoreTests extends ChunkMetadataStoreTests {
     public void setUp() throws Exception {
         val tableStore = new InMemoryTableStore(Executors.newScheduledThreadPool(10));
         metadataStore = new TableBasedMetadataStore("TEST", tableStore);
+    }
+
+    /**
+     * Unit tests for {@link TableBasedMetadataStore} with {@link InMemoryChunkStorage} using {@link SimpleStorageTests}.
+     */
+    public static class TableBasedMetadataSimpleStorageTests extends SimpleStorageTests {
+
+        protected ChunkStorage getChunkStorage() throws Exception {
+            return new InMemoryChunkStorage();
+        }
+
+        protected ChunkMetadataStore getMetadataStore() throws Exception {
+            TableStore tableStore = new InMemoryTableStore(executorService());
+            String tableName = "TableBasedMetadataSimpleStorageTests";
+            return new TableBasedMetadataStore(tableName, tableStore);
+        }
+
+        @Override
+        protected ChunkMetadataStore getCloneMetadataStore(ChunkMetadataStore metadataStore) throws Exception {
+            TableBasedMetadataStore tableBasedMetadataStore = (TableBasedMetadataStore) metadataStore;
+            TableStore tableStore = InMemoryTableStore.clone((InMemoryTableStore) tableBasedMetadataStore.getTableStore());
+            String tableName =  tableBasedMetadataStore.getTableName();
+            val retValue = new TableBasedMetadataStore(tableName, tableStore);
+            TableBasedMetadataStore.copyVersion(tableBasedMetadataStore, retValue);
+            return retValue;
+        }
+    }
+
+    /**
+     * Unit tests for {@link TableBasedMetadataStore} with {@link InMemoryChunkStorage} using {@link ChunkedRollingStorageTests}.
+     */
+    public static class InMemorySimpleStorageRollingTests extends ChunkedRollingStorageTests {
+        protected ChunkStorage getChunkStorage() throws Exception {
+            return new InMemoryChunkStorage();
+        }
+
+        protected ChunkMetadataStore getMetadataStore() throws Exception {
+            TableStore tableStore = new InMemoryTableStore(executorService());
+            String tableName = "TableBasedMetadataSimpleStorageTests";
+            return new TableBasedMetadataStore(tableName, tableStore);
+        }
+    }
+
+    /**
+     * Unit tests for {@link TableBasedMetadataStore} with {@link InMemoryChunkStorage} using {@link ChunkedSegmentStorageTests}.
+     */
+    public static class TableBasedMetadataChunkedSegmentStorageTests extends ChunkedSegmentStorageTests {
+        @Override
+        public ChunkMetadataStore createMetadataStore() throws Exception {
+            TableStore tableStore = new InMemoryTableStore(Executors.newScheduledThreadPool(1));
+            String tableName = "TableBasedMetadataSimpleStorageTests";
+            return new TableBasedMetadataStore(tableName, tableStore);
+        }
+
+        public TestContext getTestContext() throws Exception {
+            return new TableBasedMetadataTestContext(executorService());
+        }
+
+        public static class TableBasedMetadataTestContext extends TestContext {
+            TableBasedMetadataTestContext() {
+            }
+
+            TableBasedMetadataTestContext(ExecutorService executorService) throws Exception {
+                super(executorService);
+            }
+
+            @Override
+            public ChunkMetadataStore createMetadataStore() throws Exception {
+                return createChunkMetadataStore();
+            }
+
+            @Override
+            protected TestContext createNewInstance() {
+                return new TableBasedMetadataTestContext();
+            }
+
+            @Override
+            public ChunkMetadataStore getForkedMetadataStore() {
+                val thisMetadataStore = (TableBasedMetadataStore) this.metadataStore;
+                TableStore tableStore = InMemoryTableStore.clone((InMemoryTableStore) thisMetadataStore.getTableStore());
+                String tableName =  thisMetadataStore.getTableName();
+                val retValue = new TableBasedMetadataStore(tableName, tableStore);
+                TableBasedMetadataStore.copyVersion(thisMetadataStore, retValue);
+                return retValue;
+            }
+
+            private ChunkMetadataStore createChunkMetadataStore() {
+                TableStore tableStore = new InMemoryTableStore(executor);
+                String tableName = "TableBasedMetadataSimpleStorageTests";
+                return new TableBasedMetadataStore(tableName, tableStore);
+            }
+        }
     }
 }
